deleted:    .vs/MutBot/v16/.suo 	deleted:    .vs/MutBot/v16/Browse.VC.db 	deleted:    .vs/ProjectSettings.json 	deleted:    .vs/slnx.sqlite 	deleted:    CQAI/CQPdemo/CQP.lib.back 	modified:   CQAI/CQPdemo/CQPDemo.vcxproj 	modified:   CQAI/CQPdemo/CQPDemo.vcxproj.filters 	modified:   CQAI/CQPdemo/Lex.cpp 	modified:   CQAI/CQPdemo/app.json 	modified:   CQAI/CQPdemo/appmain.cpp 	modified:   CQAI/CQPdemo/lex.h 	modified:   CQAI/CQPdemo/stdafx.h 	modified:   CQAI/cn.17ds8.AItest.sln 	modified:   CQAI/notes/MAIN.md 	modified:   PyDeamon/ChineseCalendar.py 	modified:   PyDeamon/PyDeamon.py 	deleted:    PyDeamon/PyDeamon.pyproj 	deleted:    PyDeamon/PyDeamon.sln 	modified:   PyDeamon/scheduleCast.py 	modified:   PyDeamon/tmpTest.py

+++ CQAI/CQPdemo/lex.h
@@ -6,7 +6,7 @@ struct CQcmd {
 	std::int8_t		flag     = -1;
 	std::int64_t	toPri    = -1;
 	std::int64_t	toGrp    = -1;
-	std::int32_t    action   = -1;
+	std::int64_t    action   = -1;
 	std::int8_t		status   = -1;
 	const char* content = "Err";
 };
